use shared binary detection for open and reload
